requirements.txt
beautifulsoup4==4.14.*
html-to-markdown==3.3.*
marko==2.2.*
pydantic==2.13.*
PyYAML==6.0.*
requests==2.33.*